final push

+++ java/family-reader-server/src/main/java/com/techelevator/controller/FamilyController.java
@@ -72,12 +72,12 @@ public class FamilyController {
     }
 
     @RequestMapping(path="/family/{familyId}/moneyEarned", method = RequestMethod.GET)
-    public int getTotalFamilyMoneyEarned(@PathVariable ("familyId") int familyId) {
+    public double getTotalFamilyMoneyEarned(@PathVariable ("familyId") int familyId) {
         return familyDao.getFamilyMoneyEarned(familyId);
     }
 
     @RequestMapping(path="/family/{familyId}/moneyDonated", method = RequestMethod.GET)
-    public int getTotalFamilyMoneyDonated(@PathVariable ("familyId") int familyId) {
+    public double getTotalFamilyMoneyDonated(@PathVariable ("familyId") int familyId) {
         return familyDao.getFamilyMoneyDonated(familyId);
     }
